Fix resource loading issues
Fix issues with resource loading. Class.getResource assumes paths are
relative to the package unless they begin with a forward slash. All the
resources we care about are not relative to the package, so I added
forward slashes everywhere.
Fix a bug with not being able to place a card on a column.
Delete the *.class files from the repository. These should not be
committed as they are generated from source files every build.